Query Race

Query full results for a race.

query_race(result)

Arguments

result

a list or data frame with fields/columns race_id and sector that describe a single race. The easiest way to create such a data frame is through the function query_results(). This function can return multiple results, but query_race() only returns the results for one race. If multiple results are passed, only the first one will be used.

Value

A tibble with at least the following columns: rank, bib, fis_code, name, birth_year, and nation. Depending on the type of race, there are additional columns like time, run1, run2, total_time, diff_time, fis_points, and cup_points.

Details

Different types of races may have very different way to display the results. Some disciplines use time measurements, other use a points system or even a combination of different systems. In some disciplines, races involve a single run and a single time measurement, while other use multiple runs and accordingly have multiple run times and possibly a total time. The function tries to be flexible in determining the format that is used for a given race, but it is known to fail for some special cases (e.g., team races in alpine skiing).

Examples

if (FALSE) { # \dontrun{
# the results for a race can be queried by using a specific race of an
# athlete as input. So we get all downhill results for Marco Odermatt.
odermatt <- query_athletes("odermatt", "marco")
odermatt_res <- query_results(odermatt, discipline = "DH")

# show the first of the results
odermatt_res[1, ]

# get the full results for this race
query_race(odermatt_res[1, ])
} # }